+++ decidim-bulletin_board-app/app/commands/process_key_ceremony_step.rb
@@ -36,7 +36,7 @@ class ProcessKeyCeremonyStep < Rectify::Command
       log_entry.election = election
       log_entry.save!
       create_response_log_entry!
-      election.save!
+      save_election!
     end
 
     broadcast(:ok)
@@ -44,16 +44,24 @@ class ProcessKeyCeremonyStep < Rectify::Command
 
   private
 
-  attr_accessor :trustee
+  attr_accessor :trustee, :response_log_entry
 
   def create_response_log_entry!
     return unless response_message
 
-    LogEntry.create!(
+    @response_log_entry = LogEntry.create!(
       election: election,
       message_id: response_message["message_id"],
       signed_data: BulletinBoard.sign(response_message),
       bulletin_board: true
     )
+  end
+
+  def save_election!
+    if response_log_entry && response_log_entry.message_identifier.type == "end_key_ceremony"
+      election.key_ceremony_ended!
+    else
+      election.save!
+    end
   end
 end