+++ decidim-bulletin_board-app/spec/commands/start_key_ceremony_spec.rb
@@ -3,22 +3,32 @@
 require "rails_helper"
 require "./spec/commands/shared/log_entry_validations"
 
-RSpec.describe CloseBallotBox do
+RSpec.describe StartKeyCeremony do
   subject { described_class.call(client, message_id, signed_data) }
 
   include_context "with a signed message"
 
   let!(:election) { create(:election, election_status) }
-  let(:election_status) { :vote }
+  let(:election_status) { :created }
   let(:client) { Authority.first }
-  let(:message_type) { :close_ballot_box_message }
+  let(:message_type) { :start_key_ceremony_message }
   let(:message_params) { { election: election } }
 
-  it "broadcasts ok" do
-    expect { subject }.to broadcast(:ok, election)
+  shared_examples "starting the key ceremony fails" do
+    it "doesn't create a log entry" do
+      expect { subject }.not_to change(LogEntry, :count)
+    end
+
+    it "desn't change the election status" do
+      expect { subject }.not_to(change { Election.last.status })
+    end
   end
 
-  it "creates the log entry for the ballot box closing" do
+  it "broadcasts ok" do
+    expect { subject }.to broadcast(:ok)
+  end
+
+  it "creates the log entry for the message" do
     expect { subject }.to change(LogEntry, :count).by(1)
   end
 
@@ -26,30 +36,16 @@ RSpec.describe CloseBallotBox do
     expect { subject }.not_to(change(election, :voting_scheme_state))
   end
 
-  it "changes the election status" do
-    expect { subject }.to change { Election.last.status } .from("vote") .to("vote_ended")
+  it "changes the election status to key_ceremony" do
+    expect { subject }.to change { Election.last.status } .from("created").to("key_ceremony")
   end
 
-  shared_examples "closing the ballot box fails" do
-    it "doesn't create a log entry" do
-      expect { subject }.not_to change(LogEntry, :count)
-    end
+  it_behaves_like "with an invalid signed data", "starting the key ceremony fails"
 
-    it "doesn't change the voting scheme state" do
-      expect { subject }.not_to change(election, :voting_scheme_state)
-    end
+  context "when the election status is not created" do
+    let(:election_status) { :key_ceremony }
 
-    it "doesn't change the election status" do
-      expect { subject }.not_to change(election, :status)
-    end
-  end
-
-  it_behaves_like "with an invalid signed data", "closing the ballot box fails"
-
-  context "when the election status is not vote" do
-    let(:election_status) { :ready }
-
-    it_behaves_like "closing the ballot box fails"
+    it_behaves_like "starting the key ceremony fails"
 
     it "broadcasts invalid" do
       expect { subject }.to broadcast(:invalid, "The election is not in the right step")
@@ -60,9 +56,9 @@ RSpec.describe CloseBallotBox do
     let(:client) { create(:authority, private_key: private_key) }
     let(:private_key) { generate(:private_key) }
 
-    it_behaves_like "closing the ballot box fails"
+    it_behaves_like "starting the key ceremony fails"
 
-    it "broadcast invalid" do
+    it "broadcasts invalid" do
       expect { subject }.to broadcast(:invalid, "Invalid client")
     end
   end
@@ -71,9 +67,9 @@ RSpec.describe CloseBallotBox do
     let(:client) { Trustee.first }
     let(:private_key) { Test::PrivateKeys.trustees_private_keys.first }
 
-    it_behaves_like "closing the ballot box fails"
+    it_behaves_like "starting the key ceremony fails"
 
-    it "broadcast invalid" do
+    it "broadcasts invalid" do
       expect { subject }.to broadcast(:invalid, "Invalid client")
     end
   end
@@ -81,9 +77,9 @@ RSpec.describe CloseBallotBox do
   context "when the message author is not the authority" do
     let(:extra_message_params) { { authority: create(:authority) } }
 
-    it_behaves_like "closing the ballot box fails"
+    it_behaves_like "starting the key ceremony fails"
 
-    it "broadcast invalid" do
+    it "broadcasts invalid" do
       expect { subject }.to broadcast(:invalid, "Invalid message author")
     end
   end