+++ decidim-bulletin_board-app/spec/jobs/end_vote_job_spec.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+
+require "rails_helper"
+
+RSpec.describe EndVoteJob do
+  subject { described_class.perform_now(pending_message.id) }
+
+  let!(:pending_message) { create(:pending_message, election: election, message: message) }
+  let(:election) { create(:election, :vote) }
+  let(:message) { build(:end_vote_message, election: election) }
+
+  it "processes the message" do
+    expect { subject }.to change { PendingMessage.last.status } .from("enqueued").to("accepted")
+  end
+
+  context "when the message was already processed" do
+    before { described_class.perform_now(pending_message.id) }
+
+    it "doesn't change the message status" do
+      expect { subject }.not_to change { PendingMessage.last.status } .from("accepted")
+    end
+  end
+end