+++ decidim-bulletin_board-app/spec/mutations/start_vote_mutation_spec.rb
@@ -3,15 +3,22 @@
 require "rails_helper"
 
 module Mutations
-  RSpec.describe OpenBallotBoxMutation, type: :request do
+  RSpec.describe StartVoteMutation, type: :request do
     subject { post "/api", params: { query: query, variables: { messageId: message_id, signedData: signed_data } }, headers: headers }
 
     let(:query) do
       <<~GQL
         mutation($messageId: String!, $signedData: String!) {
-          openBallotBox(messageId: $messageId, signedData: $signedData) {
-            election {
+          startVote(messageId: $messageId, signedData: $signedData) {
+            pendingMessage {
               id
+              client {
+                id
+              }
+              election {
+                id
+              }
+              signedData
               status
             }
             error
@@ -20,27 +27,33 @@ module Mutations
       GQL
     end
 
-    let!(:election) { create(:election, :ready) }
+    let!(:election) { create(:election, :key_ceremony_ended) }
     let(:authority) { Authority.first }
     let(:headers) { { "Authorization": authority.api_key } }
     let(:signed_data) { JWT.encode(payload.as_json, Test::PrivateKeys.authority_private_key.keypair, "RS256") }
-    let(:payload) { build(:open_ballot_box_message, election: election) }
+    let(:payload) { build(:start_vote_message, election: election) }
     let(:message_id) { payload["message_id"] }
 
-    it "changes the election status" do
-      expect { subject }.to change { Election.last.status } .from("ready") .to("vote")
+    it "adds the message to the pending messages table" do
+      expect { subject }.to change(PendingMessage, :count).by(1)
     end
 
-    it "returns an election status" do
+    it "returns a pending message" do
       subject
-
       json = JSON.parse(response.body, symbolize_names: true)
-      data = json.dig(:data, :openBallotBox)
+      data = json.dig(:data, :startVote)
 
       expect(data).to include(
-        election: {
-          id: election.unique_id,
-          status: "vote"
+        pendingMessage: {
+          id: be_present,
+          client: {
+            id: authority.unique_id
+          },
+          election: {
+            id: election.unique_id
+          },
+          signedData: signed_data,
+          status: "enqueued"
         },
         error: nil
       )
@@ -49,17 +62,17 @@ module Mutations
     context "when the authority is not authorized" do
       let(:headers) { {} }
 
-      it "doesn't open the ballot box" do
-        expect { subject }.not_to change(election, :status)
+      it "doesn't create a pending message" do
+        expect { subject }.not_to change(PendingMessage, :count)
       end
 
       it "returns an error message" do
         subject
         json = JSON.parse(response.body, symbolize_names: true)
-        data = json.dig(:data, :openBallotBox)
+        data = json.dig(:data, :startVote)
 
         expect(data).to include(
-          election: nil,
+          pendingMessage: nil,
           error: "Authority not found"
         )
       end